Move all fixtures to pytest plugin folder
