+++ redbot/pytest/alias.py
@@ -1,0 +1,11 @@
+import pytest
+
+from redbot.cogs.alias import Alias
+from redbot.core import Config
+
+
+@pytest.fixture()
+def alias(config, monkeypatch):
+    with monkeypatch.context() as m:
+        m.setattr(Config, "get_conf", lambda *args, **kwargs: config)
+        return Alias(None)
