+++ redbot/pytest/cog_manager.py
@@ -1,0 +1,11 @@
+import pytest
+
+
+@pytest.fixture()
+def cog_mgr(red):
+    return red.cog_mgr
+
+
+@pytest.fixture()
+def default_dir(red):
+    return red.main_dir
