+++ tests/cogs/admin/test_admin.py
@@ -2,20 +2,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from redbot.cogs.admin import Admin
-from redbot.cogs.admin.announcer import Announcer
-
-
-@pytest.fixture()
-def admin(config):
-    return Admin(config)
-
-
-@pytest.fixture()
-def announcer(admin):
-    a = Announcer(MagicMock(), "Some message", admin.conf)
-    yield a
-    a.cancel()
+from redbot.pytest.admin import *
 
 
 @pytest.mark.asyncio
