+++ tests/cogs/test_alias.py
@@ -1,14 +1,5 @@
 import pytest
-
-from redbot.cogs.alias import Alias
-from redbot.core import Config
-
-
-@pytest.fixture()
-def alias(config, monkeypatch):
-    with monkeypatch.context() as m:
-        m.setattr(Config, "get_conf", lambda *args, **kwargs: config)
-        return Alias(None)
+from redbot.pytest.alias import *
 
 
 def test_is_valid_alias_name(alias):
