+++ tests/cogs/test_mod.py
@@ -1,16 +1,6 @@
 import pytest
 
-
-@pytest.fixture
-def mod(config, monkeypatch):
-    from redbot.core import Config
-
-    with monkeypatch.context() as m:
-        m.setattr(Config, "get_conf", lambda *args, **kwargs: config)
-        from redbot.core import modlog
-
-        modlog._register_defaults()
-        return modlog
+from redbot.pytest.mod import *
 
 
 @pytest.mark.asyncio
