+++ tests/core/test_cog_manager.py
@@ -2,17 +2,8 @@ from pathlib import Path
 
 import pytest
 
+from redbot.pytest.cog_manager import *
 from redbot.core import cog_manager
-
-
-@pytest.fixture()
-def cog_mgr(red):
-    return red.cog_mgr
-
-
-@pytest.fixture()
-def default_dir(red):
-    return red.main_dir
 
 
 @pytest.mark.skip
